Make SDKConfigurationService a proper core module
