+++ src/components/utils/onfidoApi.ts
@@ -3,7 +3,6 @@ import { parseJwt } from './jwt'
 import { performHttpRequest, HttpRequestParams } from '~core/Network'
 import { forEach } from './object'
 import { sendEvent, trackException } from '../../Tracker'
-import detectSystem from './detectSystem'
 
 import type {
   ImageQualityValidationPayload,
@@ -18,7 +17,6 @@ import type {
   CreateV4DocumentResponse,
   SuccessCallback,
   ErrorCallback,
-  SdkConfiguration,
   ApplicantConsent,
   ApplicantConsentStatus,
   PoASupportedCountry,
@@ -663,39 +661,3 @@ export const sendAnalytics = (
     }
   )
 }
-
-export const getSdkConfiguration = (
-  url: string,
-  token: string
-): Promise<SdkConfiguration> =>
-  new Promise((resolve, reject) => {
-    try {
-      const browserInfo = detectSystem('browser')
-      const osInfo = detectSystem('os')
-
-      const requestParams: HttpRequestParams = {
-        endpoint: `${url}/v3.3/sdk/configurations`,
-        token: `Bearer ${token}`,
-        contentType: 'application/json',
-        method: 'POST',
-        payload: JSON.stringify({
-          sdk_source: process.env.SDK_SOURCE,
-          sdk_version: process.env.SDK_VERSION,
-          sdk_metadata: {
-            system: {
-              browser: browserInfo.name,
-              browser_version: browserInfo.version,
-              os: osInfo.name,
-              os_version: osInfo.version,
-            },
-          },
-        }),
-      }
-
-      performHttpRequest(requestParams, resolve, (request) =>
-        formatError(request, reject)
-      )
-    } catch (error) {
-      reject(error)
-    }
-  })
